update to java 8

+++ src/main/java/log/analyzer/LogComparer.java
@@ -242,13 +242,13 @@ public class LogComparer {
 	public double max(ArrayList<Integer> list)
 	{
 		if (list==null) return 0.0;
-		return Collections.max(list);
+		return (double) Collections.max(list);
 	}
 	public double min(ArrayList<Integer> list)
 	{
 		if (list==null) return 0.0;
 		
-		return Collections.min(list);
+		return (double) Collections.min(list);
 	}
 	public double dev(ArrayList<Integer> list)
 	{
